Task 03_b/requirements__.txt
Flask==3.0.0
google-cloud-storage==2.15.0
